backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.0
httpx==0.27.0
anthropic==0.24.0
supabase==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openpyxl==3.1.2
weasyprint==62.3
jinja2==3.1.3
pillow==10.3.0
python-dotenv==1.0.0
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
